requirements.txt
aiohttp==3.6.2
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
bson==0.5.8
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
dataclasses==0.6
Django==2.2.10
django-grip==3.0.0
djongo==1.3.1
gripcontrol==4.0.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
kafka-python==2.0.1
multidict==4.7.5
pubcontrol==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pymongo==3.10.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
ryurest==0.1
service-identity==18.1.0
six==1.14.0
sqlparse==0.2.4
Twisted==19.10.0
txaio==20.1.1
urllib3==1.25.8
Werkzeug==0.16.1
yarl==1.4.2
zope.interface==4.7.1
